Update middleware handler to use OOP approach instead of mixin
